+++ yews/datasets/__init__.py
@@ -7,6 +7,7 @@ from .dirs import DirDataset
 from .files import DatasetArray
 from .files import FileDataset
 from .utils import *
+from .wenchuan import Wenchuan
 
 __all__ = (
     'BaseDataset',
@@ -16,4 +17,5 @@ __all__ = (
     'DatasetArray',
     'DatasetFolder',
     'DatasetArrayFolder',
+    'Wenchuan',
 )